Define ABI and ALTABI for aarch64c and riscv64c.
We considered "aarch64:64:{c,cheri}", "riscv:64:{c,cheri}:hf" and
"riscv:64:hf:{c,cheri}" for ALTABIs. We decided to use the ":cheri"
segment instead of ":c" as the choice of CheriABI is fundamental and
should be emphasised. Additionaly, the ":cheri" segment should be
placed before a floating-point ABI as it has larger impact on a
resulting ABI.

Discussed with:	brooks jhb jrtc27