Analyse if a binary is built for CheriBSD.
A binary built for CheriBSD includes two ELF notes:
* An ELF note with a tag NT_FREEBSD_ABI_TAG and a vendor FreeBSD;
* An ELF note with a tag NT_CHERIBSD_ABI_TAG and a vendor CheriBSD.

In case pkg can find these notes in a binary, a FreeBSD's vendor name is
replaced with a CheriBSD's vendor name and a FreeBSD's version is
replaced with a CheriBSD ABI version from the note's description.

Major and minor versions are left intact as they seem not to be relevant
to CheriBSD packages and might be useful to determine a FreeBSD version
the CheriBSD base system of pkg was based on.