Correctly recalculate a base Blk pointer.
picosat incorrectly assumes structure padding in the Blk structure
assuming that the Blk.data field starts BLK_FILL_BYTES (8) bytes from
the end of the structure. While it is the case on 64-bit architectures,
it is not the case on architecture with a pointer size different from
BLK_FILL_BYTES.

For example, on Morello the Blk.data field starts 16 bytes from the end
of the Blk structure because the Blk structure size is equal to 48 and
is aligned to the pointer size on Morello - 16 bytes.

Instead of subtracting the number of bytes to be allocated at the end of
the Blk structure, subtract an offest of the Blk.data field to
recalculate the base Blk address.