Don't process an invalid ABI string
Stop processing an ABI string if a semicolon cannot be found, i.e.
a number of semicolons is lower than required.

This includes the case when the string doesn't include any semicolon,
which resulted, before this change, in calling strchr(1, ':') and
a capability exception on a CHERI-enabled architecture when pkg is
compiled for CheriABI.