Add support for the benchmark ABI
Analyse ELF notes if they include a note with the name CHERI of the type
NT_CHERI_MORELLO_PURECAP_BENCHMARK_ABI. If such a note includes
description data equal to 0x1 and the binary is identified as compiled
for CheriABI, then pkg will use a package repository for the benchmark
ABI.

Since FreeBSD, CheriBSD and CHERI ELF notes can be stored in separate
sections, the code analysing the notes must search for them and update
the state describing an ABI in use but not fail if it does not find some
note.

The benchmark ABI for Arm Morello has the pkg's ABI string set to
aarch64cb and ALTABI set to aarch64:64:cheri:benchmark.